refactor(web): album lists
